backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-community
langchain_openai
langchain_pinecone
python-docx
pypdf